create myobs search screen and give myobs its own stack navigator
