+++ src/navigation/types.ts
@@ -207,6 +207,14 @@ export type ExploreStackParamList = {
   AdvancedSearch: undefined;
 };
 
+// Screens hosted by MyObservationsStackNavigator
+// The type containing the mapping must be a type alias. It cannot be an interface.
+// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+export type MyObservationsStackParamList = {
+  MyObservationsResults: undefined;
+  SearchMyObservations: undefined;
+};
+
 // Tab-only routes (not from SharedStackScreens). Intersected with SharedStackParamList
 // so TabStackParamList matches TabStackNavigator + SharedStackScreens.
 // Note from the documentation:
@@ -482,6 +490,14 @@ export type ExploreStackScreenProps<T extends keyof ExploreStackParamList> =
   CompositeScreenProps<
     NativeStackScreenProps<ExploreStackParamList, T>,
     TabStackScreenProps<"RootExplore">
+  >;
+
+// MyObservationsStackNavigator is nested inside ObsList. This composite type
+// acknowledges MyObservations screens access to outer-stack routes
+export type MyObservationsStackScreenProps<T extends keyof MyObservationsStackParamList> =
+  CompositeScreenProps<
+    NativeStackScreenProps<MyObservationsStackParamList, T>,
+    TabStackScreenProps<"ObsList">
   >;
 
 export type NoBottomTabStackScreenProps<T extends keyof NoBottomTabStackParamList> =
